src/requirements.txt
# Core ML packages
torch==2.2.2
torchvision==0.17.2
numpy<2.0.0 # Ensure compatibility with other facenet-pytorch dependencies  
pillow>=10.0.0
pandas>=2.3.3

# Face detection
facenet-pytorch>=2.6.0
opencv-python<4.9.0

# Emotion recognition
hsemotion>=0.2.5

# Utilities
kaggle>=1.5.12
python-dotenv>=1.2.1
mediapipe>=0.9.1
timm==0.6.13
tqdm>=4.66.1
pyarrow>=21.0.0
matplotlib>=1.15.0